corrections

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,19 +9,19 @@
 int main(void)
 {
 	int i;
-	unsigned long long fibonacci[98];
+	unsigned long fibonacci[98];
 
 	fibonacci[0] = 1;
 	fibonacci[1] = 2;
-	printf("%llu, %llu, ", fibonacci[0], fibonacci[1]);
+	printf("%lu, %lu, ", fibonacci[0], fibonacci[1]);
 
 	for (i = 2; i < 98; i++)
 	{
 		fibonacci[i] = fibonacci[i - 1] + fibonacci[i - 2];
 		if (i == 97)
-			printf("%llu\n", fibonacci[i]);
+			printf("%lu\n", fibonacci[i]);
 		else
-			printf("%llu, ", fibonacci[i]);
+			printf("%lu, ", fibonacci[i]);
 	}
 
 	return (0);
